--- src/components/icons/firefox-icon.tsx
@@ -1,232 +1,0 @@
-import React from 'react';
-import { Svg, BoxProps } from '@stacks/ui';
-
-export const FirefoxIcon: React.FC<BoxProps> = props => (
-  <Svg {...props}>
-    <circle cx="9" cy="9.5" r="8.5" fill="white" stroke="white" />
-    <path
-      d="M16.4208 6.53533C16.0728 5.69926 15.3668 4.79655 14.8139 4.51118C15.2083 5.27362 15.4819 6.09243 15.6251 6.93851L15.6266 6.95192C14.7209 4.69771 13.1852 3.78881 11.9309 1.80964C11.8675 1.70957 11.804 1.60929 11.7422 1.50344C11.7069 1.44298 11.6786 1.38851 11.654 1.33837C11.602 1.23779 11.5619 1.1315 11.5345 1.02165C11.5346 1.01658 11.5327 1.01167 11.5294 1.00787C11.526 1.00406 11.5214 1.00162 11.5163 1.00101C11.5114 0.999662 11.5062 0.999662 11.5013 1.00101C11.5002 1.00101 11.4986 1.00287 11.4973 1.00328C11.4961 1.0037 11.4934 1.00555 11.4915 1.00638L11.4946 1.00101C9.48272 2.17712 8.80009 4.35437 8.73727 5.44341C7.9338 5.49855 7.16563 5.79415 6.53294 6.29165C6.46693 6.23583 6.39794 6.18362 6.32627 6.13525C6.14376 5.49745 6.13598 4.82245 6.30374 4.18064C5.56489 4.53799 4.90837 5.04457 4.37573 5.66831H4.37201C4.05456 5.26658 4.07688 3.94191 4.09507 3.66521C4.00116 3.70289 3.91152 3.75041 3.82764 3.80696C3.54741 4.00665 3.28543 4.23071 3.04478 4.47652C2.77055 4.75411 2.52009 5.05414 2.29602 5.37346V5.3747V5.37325C1.78105 6.10214 1.4158 6.92565 1.22134 7.79624L1.2106 7.84907C1.19551 7.91943 1.14115 8.27185 1.13185 8.3484C1.13185 8.35438 1.13061 8.35995 1.12999 8.36594C1.05988 8.72977 1.01643 9.09823 1 9.46838V9.50965C1.00176 11.512 1.75579 13.4408 3.11309 14.9149C4.47038 16.3891 6.33218 17.3012 8.33052 17.4711C10.3289 17.6411 12.3183 17.0564 13.9058 15.8326C15.4932 14.6089 16.5631 12.8352 16.904 10.862C16.9174 10.7588 16.9284 10.6567 16.9404 10.5525C17.1053 9.19112 16.9267 7.81019 16.4208 6.53533ZM7.1889 12.7951C7.22631 12.8131 7.26144 12.8325 7.29988 12.8496L7.30546 12.8531C7.26661 12.8345 7.22776 12.8152 7.1889 12.7951ZM15.6272 6.95378V6.94614L15.6286 6.9546L15.6272 6.95378Z"
-      fill="url(#paint0_linear)"
-    />
-    <path
-      d="M16.4208 6.53538C16.0728 5.69931 15.3668 4.79659 14.8139 4.51123C15.2083 5.27366 15.4819 6.09248 15.6251 6.93856V6.94619L15.6266 6.95465C16.2439 8.7185 16.1546 10.6523 15.3773 12.352C14.4593 14.3185 12.2372 16.3342 8.75897 16.2362C5.00069 16.13 1.69089 13.3457 1.07213 9.69891C0.959493 9.12385 1.07213 8.8323 1.12875 8.36495C1.05151 8.72797 1.0084 9.0974 1 9.46843V9.5097C1.00176 11.512 1.75579 13.4409 3.11309 14.915C4.47038 16.3891 6.33218 17.3013 8.33052 17.4712C10.3289 17.6411 12.3183 17.0564 13.9058 15.8327C15.4932 14.609 16.5631 12.8352 16.904 10.862C16.9174 10.7588 16.9284 10.6567 16.9404 10.5525C17.1053 9.19116 16.9267 7.81023 16.4208 6.53538Z"
-      fill="url(#paint1_radial)"
-    />
-    <path
-      d="M16.4208 6.53538C16.0728 5.69931 15.3668 4.79659 14.8139 4.51123C15.2083 5.27366 15.4819 6.09248 15.6251 6.93856V6.94619L15.6266 6.95465C16.2439 8.7185 16.1546 10.6523 15.3773 12.352C14.4593 14.3185 12.2372 16.3342 8.75897 16.2362C5.00069 16.13 1.69089 13.3457 1.07213 9.69891C0.959493 9.12385 1.07213 8.8323 1.12875 8.36495C1.05151 8.72797 1.0084 9.0974 1 9.46843V9.5097C1.00176 11.512 1.75579 13.4409 3.11309 14.915C4.47038 16.3891 6.33218 17.3013 8.33052 17.4712C10.3289 17.6411 12.3183 17.0564 13.9058 15.8327C15.4932 14.609 16.5631 12.8352 16.904 10.862C16.9174 10.7588 16.9284 10.6567 16.9404 10.5525C17.1053 9.19116 16.9267 7.81023 16.4208 6.53538Z"
-      fill="url(#paint2_radial)"
-    />
-    <path
-      d="M12.5284 7.47536C12.5458 7.48753 12.5619 7.4997 12.5782 7.51188C12.377 7.15568 12.1266 6.82956 11.8342 6.54313C9.3447 4.05721 11.1818 1.15305 11.4916 1.00552L11.4947 1.00098C9.48276 2.17709 8.80013 4.35433 8.7373 5.44337C8.83072 5.43697 8.92331 5.42913 9.01879 5.42913C9.73266 5.43044 10.4335 5.62046 11.0499 5.97987C11.6664 6.33928 12.1765 6.85527 12.5284 7.47536Z"
-      fill="url(#paint3_radial)"
-    />
-    <path
-      d="M9.0238 7.97186C9.01057 8.17077 8.30666 8.85684 8.06051 8.85684C5.78282 8.85684 5.41309 10.2325 5.41309 10.2325C5.51394 11.3908 6.32243 12.3449 7.29977 12.8496C7.34441 12.8727 7.38967 12.8936 7.43493 12.914C7.51332 12.9487 7.59172 12.9807 7.67012 13.0102C8.00553 13.1287 8.35684 13.1963 8.71235 13.2109C12.7046 13.3979 13.4779 8.44458 10.597 7.00725C11.2766 6.91896 11.9649 7.0858 12.5283 7.47542C12.1763 6.85533 11.6662 6.33935 11.0498 5.97994C10.4333 5.62052 9.7325 5.43051 9.01863 5.4292C8.92356 5.4292 8.83056 5.43704 8.73715 5.44344C7.93368 5.49858 7.16551 5.79418 6.53282 6.29168C6.65496 6.39485 6.79281 6.53268 7.08318 6.81845C7.62672 7.35307 9.0207 7.90687 9.0238 7.97186Z"
-      fill="url(#paint4_radial)"
-    />
-    <path
-      d="M9.0238 7.97186C9.01057 8.17077 8.30666 8.85684 8.06051 8.85684C5.78282 8.85684 5.41309 10.2325 5.41309 10.2325C5.51394 11.3908 6.32243 12.3449 7.29977 12.8496C7.34441 12.8727 7.38967 12.8936 7.43493 12.914C7.51332 12.9487 7.59172 12.9807 7.67012 13.0102C8.00553 13.1287 8.35684 13.1963 8.71235 13.2109C12.7046 13.3979 13.4779 8.44458 10.597 7.00725C11.2766 6.91896 11.9649 7.0858 12.5283 7.47542C12.1763 6.85533 11.6662 6.33935 11.0498 5.97994C10.4333 5.62052 9.7325 5.43051 9.01863 5.4292C8.92356 5.4292 8.83056 5.43704 8.73715 5.44344C7.93368 5.49858 7.16551 5.79418 6.53282 6.29168C6.65496 6.39485 6.79281 6.53268 7.08318 6.81845C7.62672 7.35307 9.0207 7.90687 9.0238 7.97186Z"
-      fill="url(#paint5_radial)"
-    />
-    <path
-      d="M6.15947 6.02571C6.22437 6.06698 6.27789 6.10288 6.32481 6.13528C6.1423 5.49748 6.13452 4.82248 6.30228 4.18066C5.56343 4.53802 4.90691 5.0446 4.37427 5.66834C4.41333 5.66731 5.57522 5.64647 6.15947 6.02571Z"
-      fill="url(#paint6_radial)"
-    />
-    <path
-      d="M1.07304 9.69882C1.69118 13.3456 5.0016 16.1299 8.75988 16.2361C12.2381 16.3343 14.4602 14.3185 15.3782 12.3519C16.1556 10.6522 16.2449 8.71841 15.6275 6.95457V6.94693C15.6275 6.94095 15.6262 6.93744 15.6275 6.9393L15.6289 6.95271C15.9131 8.80498 14.9694 10.5995 13.4942 11.8129L13.4897 11.8232C10.6153 14.1604 7.86459 13.2333 7.30782 12.8549C7.26897 12.8364 7.23011 12.817 7.19126 12.7969C5.51538 11.9972 4.82304 10.4728 4.97143 9.16544C4.57361 9.17131 4.18263 9.06186 3.84586 8.85036C3.50908 8.63887 3.24097 8.33441 3.074 7.97386C3.51392 7.70479 4.01553 7.55267 4.53102 7.53201C5.04652 7.51135 5.55872 7.62283 6.01883 7.85584C6.96741 8.28577 8.04682 8.32813 9.02627 7.97386C9.02317 7.90886 7.62919 7.35486 7.08565 6.82045C6.79528 6.53467 6.65743 6.39705 6.53529 6.29367C6.46928 6.23785 6.40029 6.18564 6.32862 6.13727C6.28109 6.10488 6.22756 6.0698 6.16329 6.02771C5.57904 5.64846 4.41714 5.6693 4.3787 5.67034H4.37498C4.05754 5.2686 4.07986 3.94393 4.09805 3.66724C4.00414 3.70492 3.91449 3.75243 3.83062 3.80899C3.55038 4.00867 3.28841 4.23273 3.04776 4.47855C2.77255 4.75535 2.52105 5.05469 2.2959 5.37342V5.37466V5.37321C1.78093 6.1021 1.41567 6.92562 1.22122 7.79621C1.21729 7.81251 0.932709 9.05464 1.07304 9.69882Z"
-      fill="url(#paint7_radial)"
-    />
-    <path
-      d="M11.8347 6.54317C12.1271 6.82991 12.3776 7.15638 12.5787 7.51295C12.6227 7.54617 12.6638 7.57918 12.6988 7.61116C14.5148 9.28247 13.5633 11.645 13.4924 11.8132C14.9676 10.5997 15.9104 8.80522 15.6271 6.95295C14.721 4.69771 13.1853 3.78881 11.931 1.80964C11.8675 1.70957 11.8041 1.60929 11.7423 1.50344C11.707 1.44298 11.6786 1.38851 11.6541 1.33837C11.602 1.23779 11.5619 1.1315 11.5346 1.02165C11.5346 1.01658 11.5328 1.01167 11.5294 1.00787C11.5261 1.00406 11.5215 1.00162 11.5164 1.00101C11.5115 0.999662 11.5063 0.999662 11.5013 1.00101C11.5003 1.00101 11.4986 1.00287 11.4974 1.00328C11.4962 1.0037 11.4935 1.00555 11.4916 1.00638C11.1818 1.15308 9.34474 4.05725 11.8347 6.54317Z"
-      fill="url(#paint8_radial)"
-    />
-    <path
-      d="M12.6972 7.61013C12.6623 7.57815 12.6212 7.54513 12.5771 7.51191C12.5608 7.49974 12.5447 7.48756 12.5273 7.47539C11.9639 7.08577 11.2757 6.91893 10.596 7.00722C13.477 8.44537 12.704 13.3978 8.71141 13.2109C8.35591 13.1963 8.0046 13.1286 7.66918 13.0101C7.59078 12.9808 7.51239 12.9488 7.43399 12.914C7.38873 12.8933 7.34347 12.8727 7.29883 12.8496L7.30441 12.8531C7.86117 13.2325 10.6111 14.1596 13.4863 11.8214L13.4908 11.8111C13.5625 11.644 14.514 9.28082 12.6972 7.61013Z"
-      fill="url(#paint9_radial)"
-    />
-    <path
-      d="M5.41333 10.2325C5.41333 10.2325 5.78307 8.85683 8.06076 8.85683C8.3069 8.85683 9.01144 8.17077 9.02404 7.97186C8.04459 8.32613 6.96518 8.28377 6.0166 7.85384C5.5565 7.62084 5.04429 7.50936 4.5288 7.53002C4.0133 7.55068 3.5117 7.70279 3.07178 7.97186C3.23874 8.33241 3.50686 8.63687 3.84363 8.84837C4.18041 9.05986 4.57139 9.16931 4.9692 9.16345C4.82082 10.471 5.51316 11.9954 7.18903 12.7949C7.22644 12.8129 7.26157 12.8323 7.30001 12.8494C6.32185 12.3449 5.51419 11.3908 5.41333 10.2325Z"
-      fill="url(#paint10_radial)"
-    />
-    <path
-      d="M16.4208 6.53533C16.0728 5.69926 15.3668 4.79655 14.8139 4.51118C15.2083 5.27362 15.4819 6.09243 15.6251 6.93851L15.6266 6.95192C14.7209 4.69771 13.1852 3.78881 11.9309 1.80964C11.8675 1.70957 11.804 1.60929 11.7422 1.50344C11.7069 1.44298 11.6786 1.38851 11.654 1.33837C11.602 1.23779 11.5619 1.1315 11.5345 1.02165C11.5346 1.01658 11.5327 1.01167 11.5294 1.00787C11.526 1.00406 11.5214 1.00162 11.5163 1.00101C11.5114 0.999662 11.5062 0.999662 11.5013 1.00101C11.5002 1.00101 11.4986 1.00287 11.4973 1.00328C11.4961 1.0037 11.4934 1.00555 11.4915 1.00638L11.4946 1.00101C9.48272 2.17712 8.80009 4.35437 8.73727 5.44341C8.83068 5.43701 8.92327 5.42917 9.01875 5.42917C9.73262 5.43048 10.4334 5.6205 11.0499 5.97991C11.6663 6.33932 12.1765 6.8553 12.5284 7.47539C11.965 7.08577 11.2767 6.91893 10.5971 7.00722C13.478 8.44537 12.7051 13.3978 8.71247 13.2109C8.35696 13.1963 8.00565 13.1286 7.67023 13.0101C7.59184 12.9808 7.51344 12.9488 7.43505 12.914C7.38979 12.8933 7.34452 12.8727 7.29988 12.8496L7.30546 12.8531C7.26661 12.8345 7.22776 12.8152 7.1889 12.7951C7.22631 12.8131 7.26144 12.8325 7.29988 12.8496C6.32172 12.3449 5.51406 11.3908 5.4132 10.2324C5.4132 10.2324 5.78294 8.85681 8.06063 8.85681C8.30677 8.85681 9.01131 8.17074 9.02392 7.97184C9.02081 7.90684 7.62683 7.35283 7.0833 6.81842C6.79293 6.53265 6.65508 6.39502 6.53294 6.29165C6.46693 6.23583 6.39794 6.18362 6.32627 6.13525C6.14376 5.49745 6.13598 4.82245 6.30374 4.18064C5.56489 4.53799 4.90837 5.04457 4.37573 5.66831H4.37201C4.05456 5.26658 4.07688 3.94191 4.09507 3.66521C4.00116 3.70289 3.91152 3.75041 3.82764 3.80696C3.54741 4.00665 3.28543 4.23071 3.04478 4.47652C2.77055 4.75411 2.52009 5.05414 2.29602 5.37346V5.3747V5.37325C1.78105 6.10214 1.4158 6.92565 1.22134 7.79624L1.2106 7.84907C1.19551 7.91943 1.12793 8.2768 1.11821 8.35356C1.11821 8.35933 1.11821 8.34757 1.11821 8.35356C1.0564 8.72246 1.01693 9.09474 1 9.46838V9.50965C1.00176 11.512 1.75579 13.4408 3.11309 14.9149C4.47038 16.3891 6.33218 17.3012 8.33052 17.4711C10.3289 17.6411 12.3183 17.0564 13.9058 15.8326C15.4932 14.6089 16.5631 12.8352 16.904 10.862C16.9174 10.7588 16.9284 10.6567 16.9404 10.5525C17.1053 9.19112 16.9267 7.81019 16.4208 6.53533ZM15.6262 6.94614L15.6276 6.9546L15.6262 6.94614Z"
-      fill="url(#paint11_linear)"
-    />
-    <defs>
-      <linearGradient
-        id="paint0_linear"
-        x1="15.3606"
-        y1="3.55627"
-        x2="2.08445"
-        y2="16.386"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop offset="0.048" stopColor="#FFF44F" />
-        <stop offset="0.111" stopColor="#FFE847" />
-        <stop offset="0.225" stopColor="#FFC830" />
-        <stop offset="0.368" stopColor="#FF980E" />
-        <stop offset="0.401" stopColor="#FF8B16" />
-        <stop offset="0.462" stopColor="#FF672A" />
-        <stop offset="0.534" stopColor="#FF3647" />
-        <stop offset="0.705" stopColor="#E31587" />
-      </linearGradient>
-      <radialGradient
-        id="paint1_radial"
-        cx="0"
-        cy="0"
-        r="1"
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="translate(14.7462 2.83123) scale(16.6982 16.6712)"
-      >
-        <stop offset="0.129" stopColor="#FFBD4F" />
-        <stop offset="0.186" stopColor="#FFAC31" />
-        <stop offset="0.247" stopColor="#FF9D17" />
-        <stop offset="0.283" stopColor="#FF980E" />
-        <stop offset="0.403" stopColor="#FF563B" />
-        <stop offset="0.467" stopColor="#FF3750" />
-        <stop offset="0.71" stopColor="#F5156C" />
-        <stop offset="0.782" stopColor="#EB0878" />
-        <stop offset="0.86" stopColor="#E50080" />
-      </radialGradient>
-      <radialGradient
-        id="paint2_radial"
-        cx="0"
-        cy="0"
-        r="1"
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="translate(8.64449 9.64693) scale(16.6982 16.6712)"
-      >
-        <stop offset="0.3" stopColor="#960E18" />
-        <stop offset="0.351" stopColor="#B11927" stopOpacity="0.74" />
-        <stop offset="0.435" stopColor="#DB293D" stopOpacity="0.343" />
-        <stop offset="0.497" stopColor="#F5334B" stopOpacity="0.094" />
-        <stop offset="0.53" stopColor="#FF3750" stopOpacity="0" />
-      </radialGradient>
-      <radialGradient
-        id="paint3_radial"
-        cx="0"
-        cy="0"
-        r="1"
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="translate(10.6576 -0.952194) scale(12.0971 12.0776)"
-      >
-        <stop offset="0.132" stopColor="#FFF44F" />
-        <stop offset="0.252" stopColor="#FFDC3E" />
-        <stop offset="0.506" stopColor="#FF9D12" />
-        <stop offset="0.526" stopColor="#FF980E" />
-      </radialGradient>
-      <radialGradient
-        id="paint4_radial"
-        cx="0"
-        cy="0"
-        r="1"
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="translate(6.79737 14.0015) scale(7.95075 7.93791)"
-      >
-        <stop offset="0.353" stopColor="#3A8EE6" />
-        <stop offset="0.472" stopColor="#5C79F0" />
-        <stop offset="0.669" stopColor="#9059FF" />
-        <stop offset="1" stopColor="#C139E6" />
-      </radialGradient>
-      <radialGradient
-        id="paint5_radial"
-        cx="0"
-        cy="0"
-        r="1"
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="translate(9.55682 8.28064) rotate(-13.5704) scale(4.21507 4.9277)"
-      >
-        <stop offset="0.206" stopColor="#9059FF" stopOpacity="0" />
-        <stop offset="0.278" stopColor="#8C4FF3" stopOpacity="0.064" />
-        <stop offset="0.747" stopColor="#7716A8" stopOpacity="0.45" />
-        <stop offset="0.975" stopColor="#6E008B" stopOpacity="0.6" />
-      </radialGradient>
-      <radialGradient
-        id="paint6_radial"
-        cx="0"
-        cy="0"
-        r="1"
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="translate(8.43366 2.14913) scale(5.71976 5.71053)"
-      >
-        <stop stopColor="#FFE226" />
-        <stop offset="0.121" stopColor="#FFDB27" />
-        <stop offset="0.295" stopColor="#FFC82A" />
-        <stop offset="0.502" stopColor="#FFA930" />
-        <stop offset="0.732" stopColor="#FF7E37" />
-        <stop offset="0.792" stopColor="#FF7139" />
-      </radialGradient>
-      <radialGradient
-        id="paint7_radial"
-        cx="0"
-        cy="0"
-        r="1"
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="translate(12.9294 -1.47307) scale(24.4037 24.3642)"
-      >
-        <stop offset="0.113" stopColor="#FFF44F" />
-        <stop offset="0.456" stopColor="#FF980E" />
-        <stop offset="0.622" stopColor="#FF5634" />
-        <stop offset="0.716" stopColor="#FF3647" />
-        <stop offset="0.904" stopColor="#E31587" />
-      </radialGradient>
-      <radialGradient
-        id="paint8_radial"
-        cx="0"
-        cy="0"
-        r="1"
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="translate(10.8728 -0.118248) rotate(83.9663) scale(17.8575 11.7382)"
-      >
-        <stop stopColor="#FFF44F" />
-        <stop offset="0.06" stopColor="#FFE847" />
-        <stop offset="0.168" stopColor="#FFC830" />
-        <stop offset="0.304" stopColor="#FF980E" />
-        <stop offset="0.356" stopColor="#FF8B16" />
-        <stop offset="0.455" stopColor="#FF672A" />
-        <stop offset="0.57" stopColor="#FF3647" />
-        <stop offset="0.737" stopColor="#E31587" />
-      </radialGradient>
-      <radialGradient
-        id="paint9_radial"
-        cx="0"
-        cy="0"
-        r="1"
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="translate(8.29873 4.26424) scale(15.2356 15.211)"
-      >
-        <stop offset="0.137" stopColor="#FFF44F" />
-        <stop offset="0.48" stopColor="#FF980E" />
-        <stop offset="0.592" stopColor="#FF5634" />
-        <stop offset="0.655" stopColor="#FF3647" />
-        <stop offset="0.904" stopColor="#E31587" />
-      </radialGradient>
-      <radialGradient
-        id="paint10_radial"
-        cx="0"
-        cy="0"
-        r="1"
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="translate(12.3208 5.1545) scale(16.6753 16.6483)"
-      >
-        <stop offset="0.094" stopColor="#FFF44F" />
-        <stop offset="0.231" stopColor="#FFE141" />
-        <stop offset="0.509" stopColor="#FFAF1E" />
-        <stop offset="0.626" stopColor="#FF980E" />
-      </radialGradient>
-      <linearGradient
-        id="paint11_linear"
-        x1="15.2008"
-        y1="3.48776"
-        x2="3.90483"
-        y2="14.8018"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop offset="0.167" stopColor="#FFF44F" stopOpacity="0.8" />
-        <stop offset="0.266" stopColor="#FFF44F" stopOpacity="0.634" />
-        <stop offset="0.489" stopColor="#FFF44F" stopOpacity="0.217" />
-        <stop offset="0.6" stopColor="#FFF44F" stopOpacity="0" />
-      </linearGradient>
-    </defs>
-  </Svg>
-);
