+++ src/components/item-hover.tsx
@@ -2,7 +2,7 @@ import { Box, BoxProps, color, transition } from '@stacks/ui';
 import React from 'react';
 import { useFocus, useHover } from 'use-events';
 
-export function ItemHover({
+function ItemHover({
   isFocused,
   isHovered,
   ...rest
