+++ src/features/balances-list/components/fungible-assets.tsx
@@ -17,6 +17,7 @@ function FungibleAssetRow(props: FungibleAssetRowProps) {
 
 export function FungibleAssets(props: StackProps) {
   const fungibleTokens = useFungibleTokenBaseState();
+  console.log(fungibleTokens);
   const balances = useCurrentAccountBalancesUnanchoredState();
   if (!balances) return null;
