+++ src/pages/transaction-signing/components/error.tsx
@@ -19,7 +19,7 @@ function ErrorButton({ variant, ...props }: { variant?: 'secondary' } & BoxProps
   );
 }
 
-export interface ErrorMessageProps extends StackProps {
+interface ErrorMessageProps extends StackProps {
   title: string;
   body: string | JSX.Element;
   actions?: {
@@ -28,7 +28,6 @@ export interface ErrorMessageProps extends StackProps {
     variant?: 'secondary';
   }[];
 }
-
 export const ErrorMessage = memo(({ title, body, actions, ...rest }: ErrorMessageProps) => {
   return (
     <Stack
